moved Search component to routes & delete redundant files
